+++ libsrc/ffdec_lib/test/com/jpexs/decompiler/flash/as3decompile/ActionScript3AssembledDecompileTest.java
@@ -28,6 +28,14 @@ public class ActionScript3AssembledDecompileTest extends ActionScript3DecompileT
     }
 
     @Test
+    public void testDecrementPrecedence() {
+        decompileMethod("assembled", "testDecrementPrecedence", "var _loc2_:int = 10;\r\n"
+                + "var _loc1_:int = 5;\r\n"
+                + "var _loc3_:* = _loc2_ & (1 << _loc1_) - 1;\r\n",
+                false);
+    }
+
+    @Test
     public void testDeobfuscatorJumpsExceptionStart() {
         decompileMethod("assembled", "testDeobfuscatorJumpsExceptionStart", "try\r\n"
                 + "{\r\n"