+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - #1114 Script search results dialogs closing on swf close
 - #1159 Regexp syntax hilight when not a regexp (only division)
 - #1227 AS3 avoid recursion (stackoverflow) caused by newfunction instruction
+- #1360 Precedence of increment/decrement operations
 
 ## [13.0.2] - 2021-02-10
 ### Changed